@@ -91,8 +91,8 @@
   },
   "dependencies": {
     "@primer/octicons-react": "^16.0.0",
-    "@types/react": "^17.0.25",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.0",
     "@types/react-router-dom": "^5.1.8",
     "@types/speakingurl": "^13.0.3",
     "@types/webextension-polyfill": "^0.12.1",
@@ -104,11 +104,11 @@
     "mdast-util-to-markdown": "^1.2.3",
     "micro-match": "^1.0.1",
     "prettier": "^3.0.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router": "5.2.1",
-    "react-router-dom": "5.3.0",
-    "react-textarea-autosize": "^8.3.2",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-router": "6.26.2",
+    "react-router-dom": "6.26.2",
+    "react-textarea-autosize": "^8.5.3",
     "serialize-error": "^9.0.0",
     "speakingurl": "14.0.1",
     "strip-indent": "^4.0.0",
